pyproject.toml
[project]
name = "dnd-scheduler-bot"
version = "0.1.0"
description = "A Discord bot for D&D scheduling with multi-campaign support."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Antigravity", email = "antigravity@example.com"},
]
dependencies = [
    "discord.py>=2.3.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.2.0",
    "pytz>=2024.1",
    "coloredlogs>=15.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
